+++ streaming/src/test/scala/org/apache/spark/streaming/CheckpointSuite.scala
@@ -39,7 +39,6 @@ import org.apache.spark.internal.config._
 import org.apache.spark.rdd.RDD
 import org.apache.spark.streaming.dstream._
 import org.apache.spark.streaming.scheduler._
-import org.apache.spark.streamingtest.LocalStreamingContext
 import org.apache.spark.util.{Clock, ManualClock, MutableURLClassLoader, ResetSystemProperties, Utils}
 
 /**
