+++ streaming/src/test/scala/org/apache/spark/streaming/DStreamClosureSuite.scala
@@ -24,7 +24,6 @@ import org.scalatest.BeforeAndAfterAll
 import org.apache.spark.{HashPartitioner, SparkContext, SparkException, SparkFunSuite}
 import org.apache.spark.rdd.RDD
 import org.apache.spark.streaming.dstream.DStream
-import org.apache.spark.streamingtest.LocalStreamingContext
 import org.apache.spark.util.ReturnStatementInClosureException
 
 /**
