+++ streaming/src/test/scala/org/apache/spark/streaming/MapWithStateSuite.scala
@@ -28,7 +28,6 @@ import org.scalatest.PrivateMethodTester._
 
 import org.apache.spark.{SparkConf, SparkContext, SparkFunSuite}
 import org.apache.spark.streaming.dstream.{DStream, InternalMapWithStateDStream, MapWithStateDStream, MapWithStateDStreamImpl}
-import org.apache.spark.streamingtest.LocalStreamingContext
 import org.apache.spark.util.{ManualClock, Utils}
 
 class MapWithStateSuite extends SparkFunSuite with LocalStreamingContext
